+++ activerecord/lib/active_record/schema_dumper.rb
@@ -34,6 +34,12 @@ module ActiveRecord
     # should not be dumped to db/schema.rb.
     cattr_accessor :excl_ignore_pattern, default: /^excl_rails_[0-9a-f]{10}$/
 
+    ##
+    # :singleton-method:
+    # Specify a custom regular expression matching unique constraints which name
+    # should not be dumped to db/schema.rb.
+    cattr_accessor :unique_ignore_pattern, default: /^uniq_rails_[0-9a-f]{10}$/
+
     class << self
       def dump(connection = ActiveRecord::Base.connection, stream = STDOUT, config = ActiveRecord::Base)
         connection.create_schema_dumper(generate_options(config)).dump(stream)
@@ -182,6 +188,7 @@ module ActiveRecord
           indexes_in_create(table, tbl)
           check_constraints_in_create(table, tbl) if @connection.supports_check_constraints?
           exclusion_constraints_in_create(table, tbl) if @connection.supports_exclusion_constraints?
+          unique_keys_in_create(table, tbl) if @connection.supports_unique_keys?
 
           tbl.puts "  end"
           tbl.puts
@@ -215,6 +222,12 @@ module ActiveRecord
             exclusion_constraint_names = exclusion_constraints.collect(&:name)
 
             indexes = indexes.reject { |index| exclusion_constraint_names.include?(index.name) }
+          end
+
+          if @connection.supports_unique_keys? && (unique_keys = @connection.unique_keys(table)).any?
+            unique_key_names = unique_keys.collect(&:name)
+
+            indexes = indexes.reject { |index| unique_key_names.include?(index.name) }
           end
 
           index_statements = indexes.map do |index|